prototype of new dialog component replacing detail
